+++ srcContent/test/autotest-element.js
@@ -2,7 +2,7 @@ window.miniTests.push.apply(window.miniTests, [
 	{
 		name:'EE() / simple',
 		exec: function() {
-			var sl = EE('span')();
+			var sl = EE('span');
 			check(sl.length, 1, "First len test");
 			var s = sl[0];
 			check(s.nodeType, 1, "First node type test");
@@ -10,7 +10,7 @@ window.miniTests.push.apply(window.miniTests, [
 			check(s.childNodes.length, 0);
 			check(s.parentElement, null);
 			
-			var sl2 = EE('span', {'@title': 'mytitle'})();
+			var sl2 = EE('span', {'@title': 'mytitle'});
 			check(sl2.length, 1, "Second len test");
 			var s2 = sl2[0];
 			check(s2.nodeType, 1, "Second node type test");
@@ -20,9 +20,31 @@ window.miniTests.push.apply(window.miniTests, [
 		}
 	},
 	{
+		name:'$().add() / function',
+		exec: function() {
+			var sl = EE('span');
+			sl.add(function(obj, index) {
+				check(obj, sl[0], true);
+				check(index, 0);
+				return 'test';
+			});
+			check(sl[0].childNodes.length, 1);
+			check(sl[0].firstChild.nodeType, 3);
+			check(sl[0].firstChild.data, 'test');
+			
+			sl.add(function(obj, index) {
+				check(obj, sl[0], true);
+				check(index, 0);
+				return [EE('br'), EE('br'), 'bar'];
+			});
+			check(sl[0].childNodes.length, 4);
+			check(sl[0].childNodes[3].nodeType, 3);
+		}
+	},
+	{
 		name:'EE() / full',
 		exec: function() {
-			var sl3 = EE('div', {'@title': '5', 'className': 'a b', $marginTop: '2px'}, 'hello')();
+			var sl3 = EE('div', {'@title': '5', 'className': 'a b', $marginTop: '2px'}, 'hello');
 			check(sl3.length, 1);
 			var s3 = sl3[0];
 			check(s3.nodeType, 1);
@@ -38,22 +60,9 @@ window.miniTests.push.apply(window.miniTests, [
 		}
 	},
 	{
-		name:'EE() / onCreate',
-		exec: function() {
-			var cap = null;
-			var sl3 = EE('div', {'@title': '5', 'className': 'a b', $marginTop: '2px'}, 'hello', function(e) {
-				cap = e;
-			});
-			var s1 = sl3();
-			check(cap, s1);
-			var s2 = sl3();
-			check(cap, s2);
-		}
-	},
-	{
 		name:'EE() / complex',
 		exec: function() {
-			var sl4 = EE('div', ['hello' , EE('b', null, 'user'), '!'])();
+			var sl4 = EE('div', ['hello' , EE('b', null, 'user'), '!']);
 			var s4 = sl4[0];
 			check(s4.nodeType, 1);
 			check(/^div$/i.test(s4.tagName));
@@ -80,7 +89,7 @@ window.miniTests.push.apply(window.miniTests, [
 	{
 		name:'$().add()',
 		exec: function() {
-			var sl = EE('span')();
+			var sl = EE('span');
 			sl.add('test');
 			check(sl[0].childNodes.length, 1);
 			check(sl[0].firstChild.nodeType, 3);
@@ -102,31 +111,9 @@ window.miniTests.push.apply(window.miniTests, [
 		}
 	},
 	{
-		name:'$().add() / function',
-		exec: function() {
-			var sl = EE('span')();
-			sl.add(function(obj, index) {
-				check(obj, sl[0], true);
-				check(index, 0);
-				return 'test';
-			});
-			check(sl[0].childNodes.length, 1);
-			check(sl[0].firstChild.nodeType, 3);
-			check(sl[0].firstChild.data, 'test');
-			
-			sl.add(function(obj, index) {
-				check(obj, sl[0], true);
-				check(index, 0);
-				return [EE('br'), EE('br'), 'bar'];
-			});
-			check(sl[0].childNodes.length, 4);
-			check(sl[0].childNodes[3].nodeType, 3);
-		}
-	},
-	{
 		name:'$().fill()',
 		exec: function() {
-			var sl = EE('span')();
+			var sl = EE('span');
 			sl.fill(EE('br'));
 			check(sl[0].childNodes.length, 1);
 			check(sl[0].childNodes[0].nodeType, 1);
@@ -141,7 +128,7 @@ window.miniTests.push.apply(window.miniTests, [
 	{
 		name:'$().replace()',
 		exec: function() {
-			var sl = EE('span')();
+			var sl = EE('span');
 			sl.fill(['foo', EE('br'), 'bar']);
 			$(sl[0].childNodes[2]).replace(EE('br'));
 			check(sl[0].childNodes.length, 3);
@@ -156,7 +143,7 @@ window.miniTests.push.apply(window.miniTests, [
 	{
 		name:'$().addFront()',
 		exec: function() {
-			var sl = EE('span')();
+			var sl = EE('span');
 			sl.fill([EE('br'), 'bar']);
 			sl.addFront('foo');
 			check(sl[0].childNodes.length, 3);
@@ -168,7 +155,7 @@ window.miniTests.push.apply(window.miniTests, [
 	{
 		name:'$().addAfter()',
 		exec: function() {
-			var sl = EE('span')();
+			var sl = EE('span');
 			sl.fill([EE('br'), 'bar']);
 			$(sl[0].childNodes[0]).addAfter('foo');
 			check(sl[0].childNodes.length, 3);
@@ -184,7 +171,7 @@ window.miniTests.push.apply(window.miniTests, [
 	{
 		name:'$().addBefore()',
 		exec: function() {
-			var sl = EE('span')();
+			var sl = EE('span');
 			sl.fill([EE('br'), 'bar']);
 			$(sl[0].childNodes[0]).addBefore('foo');
 			check(sl[0].childNodes.length, 3);
@@ -200,7 +187,7 @@ window.miniTests.push.apply(window.miniTests, [
 	{
 		name:'$().fill() / multi',
 		exec: function() {
-			var sl = $([EE('span')(), EE('span')(), EE('span')()]);
+			var sl = $([EE('span'), EE('span'), EE('span')]);
 			sl.fill('foo');
 			check(sl[0].childNodes.length, 1);
 			check(sl[1].childNodes.length, 1);
@@ -213,10 +200,14 @@ window.miniTests.push.apply(window.miniTests, [
 			check(sl[0].childNodes.length, 2);
 			check(sl[1].childNodes.length, 2);
 			check(sl[2].childNodes.length, 2);
-			sl.fill(function() { return [EE('br'), 'bar']; });
+			sl.fill([EE('br'), 'bar']);
 			check(sl[0].childNodes.length, 2);
 			check(sl[1].childNodes.length, 2);
 			check(sl[2].childNodes.length, 2);
+			sl.fill([[['x']], EE('div', [EE('br')]), 'bar']);
+			check(sl[0].childNodes.length, 3);
+			check(sl[1].childNodes.length, 3);
+			check(sl[2].childNodes.length, 3);
 		}
 	},
 	{