+++ srcContent/test/autotest-toggle.js
@@ -2,8 +2,8 @@ window.miniTests.push.apply(window.miniTests, [
 	{
 		name:'$().toggle(classNames)',
 		exec: function() {
-			var s1 = EE('div')()[0];
-			var s2 = EE('div', {'className':'a b c d'})()[0];
+			var s1 = EE('div')[0];
+			var s2 = EE('div', {'className':'a b c d'})[0];
 			var t = $([s1, s2]).toggle('b d');
 			check(s1.className, '');
 			check(s2.className, 'a c');
@@ -45,8 +45,8 @@ window.miniTests.push.apply(window.miniTests, [
 	{
 		name:'$().toggle(attributes and styles)',
 		exec: function() {
-			var s1 = EE('div', {'@title':'init', $display: 'none'})()[0];
-			var s2 = EE('div', {})()[0];
+			var s1 = EE('div', {'@title':'init', $display: 'none'})[0];
+			var s2 = EE('div', {})[0];
 			var t = $([s1, s2]).toggle({'@title': 'a', $display: 'none'}, {'@title': 'b', $display: 'block'});
 			check(s1.getAttribute('title'), 'a');
 			check(s2.getAttribute('title'), 'a');